More clang compliance

+++ src/threads/analyzer.c
@@ -23,12 +23,12 @@ static void destroy() {
 }
 
 static void *analyzer_thread_routine(struct Thread *used_thread) {
-    queue = queue_create(255, sizeof(struct CpuStats) * (unsigned long) core_count);
+    queue = queue_create(255, sizeof(struct CpuStats) * core_count);
     while (thread_is_running(used_thread)) {
         thread_time(used_thread, TRUE);
         if (prevStat != NULL) {
             struct CpuStats *current = thread_read_from_buffer(thread);
-            f32 *usage = malloc(sizeof (f32) * (unsigned long) core_count);
+            f32 *usage = malloc(sizeof (f32) * core_count);
             for (int i = 0; i < core_count; ++i) {
                 usage[i] = usage_calculator_get_usage(&prevStat[i], &current[i]);
             }
